Extracted tests for source scope and type declaration
