+++ calculator/calc/model/node.py
@@ -2,7 +2,7 @@ from abc import abstractmethod
 from typing import Protocol
 
 from calc.operators.protocol import BinaryOperator
-from thinking_types.interfaces import interface
+from thinking_reflection.interfaces import interface
 
 
 @interface
