+++ calculator/calc/operators/impl.py
@@ -1,6 +1,6 @@
 from calc.operators.protocol import BinaryOperator
 
-from thinking_types.discovery import discover
+from thinking_reflection.discovery import discover
 from thinking_injection.injectable import Injectable
