+++ test/injection/context/test_phase_assignment.py
@@ -4,7 +4,7 @@ from thinking_tests.running.start import run_current_module
 from thinking_injection.context.configurable.configurator import ContextConfigurator, specialized_configurator
 from thinking_injection.context.configurable.impl import ConfigurableContext
 from thinking_injection.context.configurable.phase import ConfigurationPhase
-from thinking_types.interfaces import interface
+from thinking_reflection.interfaces import interface
 from thinking_injection.registry.customizable.customizer import TypeRegistryCustomizer
