+++ test/injection/fixtures/class_fixtures.py
@@ -2,9 +2,9 @@ from contextlib import contextmanager
 from dataclasses import dataclass
 from typing import NamedTuple, Optional, Union
 
-from thinking_types.discovery import discover
+from thinking_reflection.discovery import discover
 from thinking_injection.injectable import Injectable
-from thinking_types.interfaces import interface
+from thinking_reflection.interfaces import interface
 from thinking_injection.lifecycle import HasLifecycle, Initializable
