+++ test/injection/fixtures/subpkg/class_subfixtures.py
@@ -1,6 +1,6 @@
-from thinking_types.discovery import discover
+from thinking_reflection.discovery import discover
 from thinking_injection.injectable import Injectable
-from thinking_types.interfaces import interface
+from thinking_reflection.interfaces import interface
 
 
 @discover
