+++ test/injection/registry/test_requirements.py
@@ -3,7 +3,7 @@ from thinking_tests.running.start import run_current_module
 
 from test.util import assert_equal_dicts
 from thinking_injection.injectable import Injectable
-from thinking_types.interfaces import interface
+from thinking_reflection.interfaces import interface
 from thinking_injection.registry.protocol import TypeRegistry
 from thinking_injection.registry.simple import SimpleRegistry
