+++ thinking_injection/context/configurable/phase.py
@@ -1,8 +1,8 @@
 from abc import ABC, abstractmethod
 
-from thinking_types.discovery import discover
+from thinking_reflection.discovery import discover
 from thinking_injection.injectable import Injectable
-from thinking_types.interfaces import interface
+from thinking_reflection.interfaces import interface
 
 
 @interface
