+++ thinking_injection/registry/customizable/customizer.py
@@ -1,6 +1,6 @@
 from typing import runtime_checkable, Protocol, Optional
 
-from thinking_types.interfaces import is_concrete
+from thinking_reflection.interfaces import is_concrete
 from thinking_injection.registry.protocol import DiscoveredTypes
 from thinking_programming.collectable import Collectable
