+++ thinking_reflection/descriptor_protocol.py
