+++ thinking_reflection/discovery.py
