+++ thinking_reflection/inspect_mate.py
